requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
pydantic-core==2.20.1